examples/crewai/pyproject.toml
[project]
name = "crewai_flow_launcher"
version = "0.1.0"
description = "Launcher for crewAI flows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.5.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "jsonschema>=4.0.0",
    "litellm>=1.80.0",
    "pydantic>=2.12.4",
    "gradio>=5.23.1",
]

[project.optional-dependencies]
ollama = [
    "langchain-ollama>=0.1.0",
    "langchain-community>=0.3.0",
]
openai = ["langchain-openai>=0.2.0"]
anthropic = ["langchain-anthropic>=0.2.0"]
all-llms = [
    "langchain-ollama>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
]

[tool.setuptools]
py-modules = ["launch_flow", "main"]
